refac: テストを整理
