+++ src/jjjexperiment/main.py
@@ -373,9 +373,7 @@ def calc(input_data: dict, test_mode=False):
             q_hs_CL_d_t = q_hs_CL_d_t,
             e_rtd_C = e_rtd_C,
             q_rtd_C = q_rtd_C,
-            q_rtd_H = q_rtd_H,
             q_max_C = q_max_C,
-            q_max_H = q_max_H,
             input_C_af_C = input_C_af_C,
             dualcompressor_C = dualcompressor_C
         )
@@ -386,8 +384,7 @@ def calc(input_data: dict, test_mode=False):
             region = region,
             climateFile = climateFile,
             E_E_fan_C_d_t = E_E_fan_C_d_t,
-            q_hs_CS_d_t = q_hs_CS_d_t,
-            q_hs_CL_d_t = q_hs_CL_d_t,
+            q_hs_C_d_t = q_hs_CS_d_t + q_hs_CL_d_t,
             V_hs_supply_d_t = V_hs_supply_d_t,
             P_rac_fan_rtd_C = P_rac_fan_rtd_C,
             simu_R_C = simu_R_C,
